More splitting out from Conversion.pm, start Parse.pm
grabbed the sub that parses the lines from Conversion and
made a Parse.pm Still need to get this tidied up so it
creates a parsed version of the Lst file.

Will then be able to code up the conversions on the parsed
lines. the extracted operation also contains the reformat
code.

This needs split off to its own routine.